+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Major features:
 - help page, listing all key-bindings
 - a job is said to be *successful* when there's no error, test failure or warning. When a job is successful, its output is displayed by bacon. This makes it possible to have a `cargo run` job, for example.
 - it's possible to define an *action* to run when a job is successful. For example you can launch a `cargo doc --open` job on a key, and have bacon switch to the previous job with the `on_success = "back` trigger so that you don't open a browser page on every change
+- arguments given after `--` are given to the job - Fix #67
 Minor changes:
 - fix character being lost behind scrollbar on wrapping